+++ src/App.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
-import LoginPage from './components/login-page';
-import Map from './components/map';
+import LoginPage from './pages/login-page';
+import Map from './pages/map-page';
+import Profile from './pages/profile-page';
 import './scss/App.scss';
-import { pages } from './consts';
+import { pages } from './constants';
 
 export default class App extends Component {
-  state = { page: pages.login };
+  state = { page: pages.login.key };
 
   setPage = (page) => {
     this.setState({ page });
@@ -16,8 +17,13 @@ export default class App extends Component {
 
     return (
       <div className='App'>
-        {page === pages.login && <LoginPage setPage={this.setPage} />}
-        {page === pages.map && <Map setPage={this.setPage} />}
+        {
+          {
+            [pages.login.key]: <LoginPage setPage={this.setPage} />,
+            [pages.map.key]: <Map setPage={this.setPage} />,
+            [pages.profile.key]: <Profile setPage={this.setPage} />
+          }[page]
+        }
       </div>
     )
   }
