+++ src/constants.js
@@ -1,0 +1,5 @@
+export const pages = {
+  map: { key: 'map', name: 'Карта' },
+  profile: { key: 'profile', name: 'Профиль' },
+  login: { key: 'login', name: 'Выйти' }
+};
