--- src/consts.js
@@ -1,5 +1,0 @@
-export const pages = {
-  login: 'login',
-  map: 'map',
-  profile: 'profile'
-};
