+++ src/scss/App.scss
@@ -7,3 +7,13 @@
   height: 100%;
   width: 100%;
 }
+
+.page-container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.page-content {
+  flex: 1;
+}
