+++ src/scss/header.scss
@@ -4,6 +4,7 @@
   justify-content: space-between;
   align-items: center;
   height: 100px;
+  min-height: 100px;
 
   &__logo {
     margin: 0 27px;
