--- src/scss/map.scss
@@ -1,9 +1,0 @@
-.map-container {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.map {
-  flex: 1;
-}
